Move debugBtn to a better spot
