+++ src/components/shared/Button/CommandBarButton.js
@@ -14,6 +14,28 @@ type Props = {
   pressed?: boolean
 };
 
+export function debugBtn(
+  onClick: ?Function,
+  type: string,
+  className: string,
+  tooltip: string,
+  disabled: boolean = false,
+  ariaPressed: boolean = false
+) {
+  return (
+    <CommandBarButton
+      className={classnames(type, className)}
+      disabled={disabled}
+      key={type}
+      onClick={onClick}
+      pressed={ariaPressed}
+      title={tooltip}
+    >
+      <img className={type} />
+    </CommandBarButton>
+  );
+}
+
 const CommandBarButton = (props: Props) => {
   const { children, className, pressed = false, ...rest } = props;
